Make ringDescriber to not rely on Session and move it to separate file
