+++ session.go
@@ -151,8 +151,14 @@ func NewSession(cfg ClusterConfig) (*Session, error) {
 
 	s.routingKeyInfoCache.lru = lru.New(cfg.MaxRoutingKeyInfo)
 
-	s.hostSource = &ringDescriber{session: s}
-	s.ringRefresher = newRefreshDebouncer(ringRefreshDebounceTime, func() error { return refreshRing(s.hostSource) })
+	s.hostSource = &ringDescriber{cfg: &s.cfg, logger: s.logger}
+	s.ringRefresher = newRefreshDebouncer(ringRefreshDebounceTime, func() error {
+		hosts, partitioner, err := s.hostSource.GetHosts()
+		if err != nil {
+			return err
+		}
+		return refreshRing(hosts, partitioner, s)
+	})
 
 	if cfg.PoolConfig.HostSelectionPolicy == nil {
 		cfg.PoolConfig.HostSelectionPolicy = RoundRobinHostPolicy()
@@ -258,6 +264,8 @@ func (s *Session) init() error {
 			s.usingTimeoutClause = " USING TIMEOUT " + strconv.FormatInt(int64(s.cfg.MetadataSchemaRequestTimeout.Milliseconds()), 10) + "ms"
 		}
 		conn.mu.Unlock()
+
+		s.hostSource.setControlConn(s.control)
 
 		if !s.cfg.DisableInitialHostLookup {
 			var partitioner string
